SWING + ORACLE (not fully implemented yet)

+++ src/View/Process/Exchanger.java
@@ -5,10 +5,10 @@ import Model.Money;
 
 public class Exchanger {
 
-    private final float quantity;
+    private final double quantity;
     private final ExchangeRate exchangeRate;
 
-    public Exchanger(float quantity, ExchangeRate exchangeRate) {
+    public Exchanger(double quantity, ExchangeRate exchangeRate) {
         this.quantity = quantity;
         this.exchangeRate = exchangeRate;
     }
@@ -17,7 +17,7 @@ public class Exchanger {
         return new Money(getExchange(), exchangeRate.getCurrencyTo());
     }
 
-    private float getExchange() {
+    private double getExchange() {
         return quantity * exchangeRate.getRate();
     }
 }
